refactor: move main attribute into separate module
